Fix Python 3 bug with integer division
The midPointAccessor variable, which is used various places in the code
(mainly but not exclusively in verbose prints), is supposed to be a
tuple of integers to index the array.  But the form used for the
division:

(tp-1)/2

produces a floating point result in Python 3.  And a tuple of floats is
no longer allowed as an index to a numpy array.  The fix is to force
integer division (which is clearly the intent here):

(tp-1)//2

This fix was made 3 places in the code.  It fixes an error in the print
statements when beVerbose=True is specified.